refactor(#1): 디렉토리 이름 복원 - ErrorProperty와 GlobalExceptionHandler가 같이 있기 때문에 error가 더 맞는 디렉토리 이름이라고 판단하여 다시 복원했어요.
