+++ src/main/java/com/teka/shared/error/GlobalErrorProperty.java
@@ -1,4 +1,4 @@
-package com.teka.shared.exception;
+package com.teka.shared.error;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
