+++ src/main/java/com/teka/shared/error/TekaException.java
@@ -1,4 +1,4 @@
-package com.teka.shared.exception;
+package com.teka.shared.error;
 
 import lombok.Getter;
